Add hook to read width of elements
This only works on the resize of the _window_ and not the resize
of the element; this is because I am using the window events and
not a resize observer.  A more permenant solution to this problem
is to instead use a resize observer with a javascript shim for
browsers that don't support it yet, and pass the information down
in a context provider, using the context hook to read the info in
any element that needs it.