Flesh out transition from index to blog post page
This commit is far to big but here goes

Add my own version of "pushd" which, when enabled links are clicked,
pushes the link onto a stack whch is displayed at the bottom to
make it easier to keep track of which pages you've visited on the
site.

Organization is quite poor at the moment, and the text-ui elements
have been broken for the moment, anticipating the addition of a
context provider that they are best used within.

The PushD components should be in their own place, and the theming
needs to be exported to a global stylesheet which perhaps uses a
context to set and track the currently selected theme.