Fix strange bug involving backgrounds and links
Backgrounds were never updating when navigating back to the index
page from a post.  The components never appeared to re-draw when
isVisible was called, backgrounds were never set.

Adding an if condition here seems to help.

This may be because of an issue with passing objects rather than
state down to the components, and the components requiring a change
to their BG props in order to re-draw.  Proper use of context must
be investigated.